Ask before deleting a subscription
Fixes #120

+++ src/bitmessageqt/__init__.py
@@ -3380,6 +3380,8 @@ class MyForm(settingsmixin.SMainWindow):
         self.click_pushButtonAddSubscription()
         
     def on_action_SubscriptionsDelete(self):
+        if QtGui.QMessageBox.question(self, "Delete subscription?", _translate("MainWindow", "If you delete the subscription, messages that you already received will become inaccessible. Maybe you can consider disabling the subscription instead. Disabled subscriptions will not receive new messages, but you can still view messages you already received.\n\nAre you sure you want to delete the subscription?"), QMessageBox.Yes|QMessageBox.No) != QMessageBox.Yes:
+            return
         address = self.getCurrentAccount()
         sqlExecute('''DELETE FROM subscriptions WHERE address=?''',
                    address)
